+++ src/scales.jl
@@ -562,7 +562,19 @@ function mergescales(c1::ContinuousScale, c2::ContinuousScale)
     c1.force && c2.force && assert_equal(c1.extrema, c2.extrema)
     i = findfirst((c1.force, c2.force))
     force = !isnothing(i)
-    extrema = force ? (c1.extrema, c2.extrema)[i] : extend_extrema(c1.extrema, c2.extrema)
+    extrema = if force
+        (c1.extrema, c2.extrema)[i]
+    else
+        result = try
+            extend_extrema(c1.extrema, c2.extrema)
+        catch err
+            sprint(Base.showerror, err)
+        end
+        if result isa String
+            error("Merging the extrema of two continuous scales failed. This happens if two layers are combined which use data of different types for the same scale.\nThe incompatible extrema of the two scales were $(c1.extrema) and $(c2.extrema).\nThe error was: $result")
+        end
+        result
+    end
     label = mergelabels(c1.label, c2.label)
     if c1.props != c2.props
         error("Expected props of merging continuous scales to match, got $(c1.props) and $(c2.props)")